requirements.txt
flask
supabase
python-dotenv
requests
gunicorn